Move class Commit to _git.py
Move _commits() to Git.commits()
